# frontend/src/generated/backend-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Starlib API",
    "description": "Backend API for Starlib music management",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/rulesets": {
      "get": {
        "tags": [
          "rulesets"
        ],
        "summary": "List Rulesets",
        "description": "Return all rulesets and the active ruleset id.",
        "operationId": "list_rulesets_api_rulesets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RulesetsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "rulesets"
        ],
        "summary": "Create Ruleset",
        "description": "Create a new user ruleset.",
        "operationId": "create_ruleset_api_rulesets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RulesetCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ruleset"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/rulesets/{ruleset_id}": {
      "put": {
        "tags": [
          "rulesets"
        ],
        "summary": "Update Ruleset",
        "description": "Update a ruleset's name and/or rules.\n\nRaises\n------\nHTTPException\n    404 if not found, 403 if built-in.",
        "operationId": "update_ruleset_api_rulesets__ruleset_id__put",
        "parameters": [
          {
            "name": "ruleset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ruleset Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RulesetUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ruleset"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "rulesets"
        ],
        "summary": "Delete Ruleset",
        "description": "Delete a user ruleset.\n\nRaises\n------\nHTTPException\n    404 if not found, 403 if built-in.",
        "operationId": "delete_ruleset_api_rulesets__ruleset_id__delete",
        "parameters": [
          {
            "name": "ruleset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ruleset Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rulesets/{ruleset_id}/activate": {
      "put": {
        "tags": [
          "rulesets"
        ],
        "summary": "Activate Ruleset",
        "description": "Set the active ruleset.\n\nRaises\n------\nHTTPException\n    404 if not found.",
        "operationId": "activate_ruleset_api_rulesets__ruleset_id__activate_put",
        "parameters": [
          {
            "name": "ruleset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ruleset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ruleset"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/folders/ruleset": {
      "get": {
        "tags": [
          "folders"
        ],
        "summary": "Get Folder Ruleset",
        "description": "Return the resolved ruleset binding for an absolute folder path.",
        "operationId": "get_folder_ruleset_api_folders_ruleset_get",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Absolute folder path",
              "title": "Path"
            },
            "description": "Absolute folder path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderRulesetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "folders"
        ],
        "summary": "Set Folder Ruleset",
        "description": "Bind a ruleset to an absolute folder path.",
        "operationId": "set_folder_ruleset_api_folders_ruleset_put",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Absolute folder path",
              "title": "Path"
            },
            "description": "Absolute folder path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderRulesetUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderRulesetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "folders"
        ],
        "summary": "Delete Folder Ruleset",
        "description": "Remove the ruleset binding for a folder path.",
        "operationId": "delete_folder_ruleset_api_folders_ruleset_delete",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Absolute folder path",
              "title": "Path"
            },
            "description": "Absolute folder path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderRulesetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FolderRulesetResponse": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "ruleset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ruleset Id"
          },
          "recursive": {
            "type": "boolean",
            "title": "Recursive",
            "default": false
          },
          "source_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Path"
          }
        },
        "type": "object",
        "required": [
          "path",
          "ruleset_id"
        ],
        "title": "FolderRulesetResponse",
        "description": "A resolved folder ruleset binding.\n\nWhen the ``source_path`` differs from the requested path, the binding\nwas inherited from a recursive ancestor."
      },
      "FolderRulesetUpdate": {
        "properties": {
          "ruleset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ruleset Id"
          },
          "recursive": {
            "type": "boolean",
            "title": "Recursive",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "ruleset_id"
        ],
        "title": "FolderRulesetUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Rule": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "enum": [
              "move",
              "convert",
              "copy"
            ],
            "title": "Type"
          },
          "input": {
            "type": "string",
            "title": "Input",
            "default": "source"
          },
          "requires": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Requires"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "Rule",
        "description": "A single step in a finalization pipeline."
      },
      "Ruleset": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "is_builtin": {
            "type": "boolean",
            "title": "Is Builtin",
            "default": false
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/Rule"
            },
            "type": "array",
            "title": "Rules"
          },
          "required_attributes": {
            "items": {
              "type": "string",
              "enum": [
                "title",
                "artist",
                "genre",
                "bpm",
                "key",
                "release_date",
                "remixer",
                "comment",
                "artwork"
              ]
            },
            "type": "array",
            "title": "Required Attributes"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Ruleset",
        "description": "A named, ordered collection of rules."
      },
      "RulesetCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/Rule"
            },
            "type": "array",
            "title": "Rules"
          },
          "required_attributes": {
            "items": {
              "type": "string",
              "enum": [
                "title",
                "artist",
                "genre",
                "bpm",
                "key",
                "release_date",
                "remixer",
                "comment",
                "artwork"
              ]
            },
            "type": "array",
            "title": "Required Attributes"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "RulesetCreate",
        "description": "Payload for creating a new ruleset."
      },
      "RulesetUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "rules": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Rule"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rules"
          },
          "required_attributes": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "enum": [
                    "title",
                    "artist",
                    "genre",
                    "bpm",
                    "key",
                    "release_date",
                    "remixer",
                    "comment",
                    "artwork"
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required Attributes"
          }
        },
        "type": "object",
        "title": "RulesetUpdate",
        "description": "Payload for updating an existing ruleset."
      },
      "RulesetsResponse": {
        "properties": {
          "rulesets": {
            "items": {
              "$ref": "#/components/schemas/Ruleset"
            },
            "type": "array",
            "title": "Rulesets"
          },
          "active_ruleset_id": {
            "type": "string",
            "title": "Active Ruleset Id"
          }
        },
        "type": "object",
        "required": [
          "rulesets",
          "active_ruleset_id"
        ],
        "title": "RulesetsResponse",
        "description": "Response listing all rulesets."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}